refactor: changed ticket price type from double to big decimal

+++ backend/src/main/java/com/example/backend/model/Reservation.java
@@ -8,6 +8,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -33,6 +35,6 @@ public class Reservation {
     @OneToMany(mappedBy = "reservation", cascade = CascadeType.ALL)
     private List<Ticket> tickets;
 
-    private double totalPrice;
+    private BigDecimal totalPrice;
 
 }
